+++ src/lib/accessibility/Behaviors/Menu/menuBehavior.ts
@@ -3,6 +3,7 @@ import { FocusZoneDirection } from '../../FocusZone'
 
 /**
  * @description
+ * Implements ARIA Menu design pattern.
  * The 'menu' role is used to identify an element that creates a list of common actions or functions that a user can invoke.
  *
  * @specification
@@ -21,7 +22,7 @@ const menuBehavior: Accessibility = (props: any) => ({
     props: {
       isCircularNavigation: true,
       preventDefaultWhenHandled: true,
-      shouldFocusFirstElementWhenReceivedFocus: true,
+      shouldFocusInnerElementWhenReceivedFocus: true,
       direction: props.vertical ? FocusZoneDirection.vertical : FocusZoneDirection.horizontal,
     },
   },